fix(license-v2): enforce integer type for limit and offset parameters
- Updated the `limit` and `offset` query parameters in the license-v2 router to enforce integer types, ensuring proper validation and preventing potential issues with non-integer values.

This change enhances the robustness of the API by ensuring that pagination parameters are correctly typed.

+++ backend/src/ee/routes/v1/license-v2-router.ts
@@ -295,8 +295,8 @@ export const registerLicenseV2Router = async (server: FastifyZodProvider) => {
       params: z.object({ organizationId: z.string().trim().uuid() }),
       querystring: z.object({
         search: z.string().trim().max(255).optional().describe("Match root organizations whose name contains this."),
-        limit: z.coerce.number().min(1).max(1000).default(100).describe("Maximum organizations to return."),
-        offset: z.coerce.number().min(0).default(0).describe("Number of organizations to skip.")
+        limit: z.coerce.number().int().min(1).max(1000).default(100).describe("Maximum organizations to return."),
+        offset: z.coerce.number().int().min(0).default(0).describe("Number of organizations to skip.")
       }),
       response: {
         200: z.object({
